The program runs. added a dedstroy for LinkedList but still not to others
